renames homePanel dir to modLauncher
