+++ src/sidebar/modLauncher/ActiveModListItem.module.scss
