+++ src/sidebar/modLauncher/ActiveModListItem.tsx
